fixed the sendWelcomeEmail function so you can now create an account, recieve an email, and then log in with that account.
