+++ public/css/style.css
@@ -876,9 +876,9 @@ box-shadow: 6px 5px var(--light-dark);
 } */
 
 
-#password-input-login {
+/* #password-input-login {
 
-}
+} */
 
 
 #username-input-signup {
