+++ views/signup.handlebars
@@ -1,26 +1,26 @@
-<div class="card">
+<div class="card" id="signup-form">
   <div class="card-header">
     <h2>Create your account</h2>
-    <h4>Create an account to post your own workouts!</h4>
   </div>
   <div id="signup" role="tabpanel">
     <form id="signup-form" class="card-body">
       <div>
-        <label for="username-input-signup">Username</label>
+        <label for="username-input-signup" class="form-label">Username</label>
         <input type="text" id="username-input-signup" />
       </div>
       <div>
-        <label for="email-input-signup">Email</label>
+        <label for="email-input-signup" class="form-label">Email</label>
         <input type="email" id="email-input-signup" />
       </div>
       <div>
-        <label for="password-input-signup">Password</label>
+        <label for="password-input-signup" class="form-label">Password</label>
         <input type="password" id="password-input-signup" />
       </div>
-      <button type="submit" id="signup-btn" class="btn">Create your account</button>
-      <a href="/login">Login instead</a>
+      <button type="submit" id="signup-btn" class="btn">Create</button>
+      <a class="form-label" href="/login">Login instead</a>
     </form>
   </div>
 </div>
+
 
 <script src="/js/signup.js"></script>
